% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{Ontology}
\alias{Ontology}
\title{Ontology}
\value{
Response object

Response object

Response object

Response object
}
\description{
Ontology object, inheriting from ConceptLibraryClient::Endpoint - allows
querying of ontology/ endpoints
}
\section{Super class}{
\code{\link[ConceptLibraryClient:Endpoint]{ConceptLibraryClient::Endpoint}} -> \code{Ontology}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Ontology-get}{\code{Ontology$get()}}
\item \href{#method-Ontology-get_detail}{\code{Ontology$get_detail()}}
\item \href{#method-Ontology-get_groups}{\code{Ontology$get_groups()}}
\item \href{#method-Ontology-get_group_detail}{\code{Ontology$get_group_detail()}}
\item \href{#method-Ontology-clone}{\code{Ontology$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ConceptLibraryClient" data-topic="Endpoint" data-id="initialize"><a href='../../ConceptLibraryClient/html/Endpoint.html#method-Endpoint-initialize'><code>ConceptLibraryClient::Endpoint$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ontology-get"></a>}}
\if{latex}{\out{\hypertarget{method-Ontology-get}{}}}
\subsection{Method \code{get()}}{
Queries ontology/node/, with optional query parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ontology$get(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(list) List of optional parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ontology-get_detail"></a>}}
\if{latex}{\out{\hypertarget{method-Ontology-get_detail}{}}}
\subsection{Method \code{get_detail()}}{
Queries ontology/node/{id}/
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ontology$get_detail(node_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{node_id}}{(string) Id of entity to query}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ontology-get_groups"></a>}}
\if{latex}{\out{\hypertarget{method-Ontology-get_groups}{}}}
\subsection{Method \code{get_groups()}}{
Queries ontology/
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ontology$get_groups()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ontology-get_group_detail"></a>}}
\if{latex}{\out{\hypertarget{method-Ontology-get_group_detail}{}}}
\subsection{Method \code{get_group_detail()}}{
Queries ontology/type/{id}/
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ontology$get_group_detail(ontology_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ontology_id}}{(string) Id of entity to query}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ontology-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Ontology-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ontology$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
